import: Reuse get_last_message_id() helper.
